modify hvm_pci_register to use uint16_t for domain instead of uint8_t

+++ tools/libxc/xc_domain.c
@@ -1653,7 +1653,7 @@ out:
 }
 
 int xc_hvm_register_pcidev(xc_interface *xch, domid_t dom, ioservid_t id,
-                           uint8_t domain, uint8_t bus, uint8_t device,
+                           uint16_t domain, uint8_t bus, uint8_t device,
                            uint8_t function)
 {
     DECLARE_HYPERCALL;
