+++ tools/libxc/xenctrl.h
@@ -1682,7 +1682,7 @@ int xc_hvm_unmap_io_range_from_ioreq_server(xc_interface *xch, domid_t dom,
  * Register a PCI device
  */
 int xc_hvm_register_pcidev(xc_interface *xch, domid_t dom, unsigned int id,
-                           uint8_t domain, uint8_t bus, uint8_t device,
+                           uint16_t domain, uint8_t bus, uint8_t device,
                            uint8_t function);
 
 /* HVM guest pass-through */
